# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceFix API",
    "description": "Investigate failed GitHub Actions runs and independently verify proposed patches.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/repositories/{repo_id}/policy": {
      "put": {
        "summary": "Put Policy",
        "operationId": "put_policy_v1_repositories__repo_id__policy_put",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Put Policy V1 Repositories  Repo Id  Policy Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repo_id}/repair-runs": {
      "post": {
        "summary": "Manual Repair Run",
        "operationId": "manual_repair_run_v1_repositories__repo_id__repair_runs_post",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualInvestigation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Manual Repair Run V1 Repositories  Repo Id  Repair Runs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ManualInvestigation": {
        "properties": {
          "github_run_id": {
            "type": "integer",
            "title": "Github Run Id"
          },
          "github_attempt": {
            "type": "integer",
            "title": "Github Attempt",
            "default": 1
          },
          "source_sha": {
            "type": "string",
            "title": "Source Sha"
          },
          "execution_sha": {
            "type": "string",
            "title": "Execution Sha"
          },
          "base_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Sha"
          }
        },
        "type": "object",
        "required": [
          "github_run_id",
          "source_sha",
          "execution_sha"
        ],
        "title": "ManualInvestigation"
      },
...
      "PolicyUpdate": {
        "properties": {
          "document": {
            "type": "object",
            "title": "Document"
          },
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "document",
          "expected_version"
        ],
        "title": "PolicyUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}